+++ Source/minecraftplus_mod/BlockHolidayLight_2.java
@@ -1,0 +1,254 @@
+package net.minecraftplus_mod;
+
+import static net.minecraftforge.common.ForgeDirection.EAST;
+import static net.minecraftforge.common.ForgeDirection.NORTH;
+import static net.minecraftforge.common.ForgeDirection.SOUTH;
+import static net.minecraftforge.common.ForgeDirection.WEST;
+
+import java.util.Random;
+
+import cpw.mods.fml.common.Side;
+import cpw.mods.fml.common.asm.SideOnly;
+import net.minecraft.src.AxisAlignedBB;
+import net.minecraft.src.IBlockAccess;
+import net.minecraft.src.Material;
+import net.minecraft.src.World;
+
+public class BlockHolidayLight_2 extends BlockNormal{
+
+	/** Whether this lamp block is the powered version. */
+    private final boolean powered;
+	
+	public BlockHolidayLight_2(int par1, boolean isPowered) {
+		super(par1, 28, Material.circuits);
+		
+		this.powered = isPowered;
+		if (isPowered)
+        {
+            this.setLightValue(0.8F);
+            ++this.blockIndexInTexture;
+        }
+	}
+
+	/**
+     * Called whenever the block is added into the world. Args: world, x, y, z
+     */
+    public void onBlockAdded(World par1World, int par2, int par3, int par4)
+    {
+        if (!par1World.isRemote)
+        {
+            if (this.powered && !par1World.isBlockIndirectlyGettingPowered(par2, par3, par4))
+            {
+                par1World.scheduleBlockUpdate(par2, par3, par4, this.blockID, 4);
+            }
+            else if (!this.powered && par1World.isBlockIndirectlyGettingPowered(par2, par3, par4))
+            {
+            	int m = par1World.getBlockMetadata(par2, par3, par4);
+                par1World.setBlockAndMetadataWithNotify(par2, par3, par4, MinecraftPlusBase.holidaylights_2Active.blockID, m);
+            }
+        }
+    }
+    /**
+     * Lets the block know when one of its neighbor changes. Doesn't know which neighbor changed (coordinates passed are
+     * their own) Args: x, y, z, neighbor blockID
+     */
+    public void onNeighborBlockChange(World par1World, int par2, int par3, int par4, int par5)
+    {
+        if (!par1World.isRemote)
+        {
+            if (this.powered && !par1World.isBlockIndirectlyGettingPowered(par2, par3, par4))
+            {
+                par1World.scheduleBlockUpdate(par2, par3, par4, this.blockID, 4);
+            }
+            else if (!this.powered && par1World.isBlockIndirectlyGettingPowered(par2, par3, par4))
+            {
+            	int m = par1World.getBlockMetadata(par2, par3, par4);
+                par1World.setBlockAndMetadataWithNotify(par2, par3, par4, MinecraftPlusBase.holidaylights_2Active.blockID, m);
+            }
+        }
+        int var6 = par1World.getBlockMetadata(par2, par3, par4);
+        boolean var7 = false;
+
+        if (var6 == 2 && par1World.isBlockSolidOnSide(par2, par3, par4 + 1, NORTH))
+        {
+            var7 = true;
+        }
+
+        if (var6 == 3 && par1World.isBlockSolidOnSide(par2, par3, par4 - 1, SOUTH))
+        {
+            var7 = true;
+        }
+
+        if (var6 == 4 && par1World.isBlockSolidOnSide(par2 + 1, par3, par4, WEST))
+        {
+            var7 = true;
+        }
+
+        if (var6 == 5 && par1World.isBlockSolidOnSide(par2 - 1, par3, par4, EAST))
+        {
+            var7 = true;
+        }
+
+        if (!var7)
+        {
+            this.dropBlockAsItem(par1World, par2, par3, par4, var6, 0);
+            par1World.setBlockWithNotify(par2, par3, par4, 0);
+        }
+
+        super.onNeighborBlockChange(par1World, par2, par3, par4, par5);
+    }
+
+    /**
+     * Ticks the block if it's been scheduled
+     */
+    public void updateTick(World par1World, int par2, int par3, int par4, Random par5Random)
+    {
+        if (!par1World.isRemote && this.powered && !par1World.isBlockIndirectlyGettingPowered(par2, par3, par4))
+        {
+        	int m = par1World.getBlockMetadata(par2, par3, par4);
+            par1World.setBlockAndMetadataWithNotify(par2, par3, par4, MinecraftPlusBase.holidaylights_2Idle.blockID, m);
+        }
+    }    
+    
+    /**
+     * Returns a bounding box from the pool of bounding boxes (this means this box can change after the pool has been
+     * cleared to be reused)
+     */
+    public AxisAlignedBB getCollisionBoundingBoxFromPool(World par1World, int par2, int par3, int par4)
+    {
+        this.setBlockBoundsBasedOnState(par1World, par2, par3, par4);
+        return super.getCollisionBoundingBoxFromPool(par1World, par2, par3, par4);
+    }
+
+    @SideOnly(Side.CLIENT)
+
+    /**
+     * Returns the bounding box of the wired rectangular prism to render.
+     */
+    public AxisAlignedBB getSelectedBoundingBoxFromPool(World par1World, int par2, int par3, int par4)
+    {
+        this.setBlockBoundsBasedOnState(par1World, par2, par3, par4);
+        return super.getSelectedBoundingBoxFromPool(par1World, par2, par3, par4);
+    }
+
+    /**
+     * Updates the blocks bounds based on its current state. Args: world, x, y, z
+     */
+    public void setBlockBoundsBasedOnState(IBlockAccess par1IBlockAccess, int par2, int par3, int par4)
+    {
+        this.func_85107_d(par1IBlockAccess.getBlockMetadata(par2, par3, par4));
+    }
+
+    public void func_85107_d(int par1)
+    {
+        float var3 = 0.125F;
+
+        if (par1 == 2)
+        {
+            this.setBlockBounds(0.0F, 0.0F, 1.0F - var3, 1.0F, 1.0F, 1.0F);
+        }
+
+        if (par1 == 3)
+        {
+            this.setBlockBounds(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, var3);
+        }
+
+        if (par1 == 4)
+        {
+            this.setBlockBounds(1.0F - var3, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F);
+        }
+
+        if (par1 == 5)
+        {
+            this.setBlockBounds(0.0F, 0.0F, 0.0F, var3, 1.0F, 1.0F);
+        }
+    }
+
+    /**
+     * Is this block (a) opaque and (b) a full 1m cube?  This determines whether or not to render the shared face of two
+     * adjacent blocks and also whether the player can attach torches, redstone wire, etc to this block.
+     */
+    public boolean isOpaqueCube()
+    {
+        return false;
+    }
+
+    /**
+     * If this block doesn't render as an ordinary block it will return False (examples: signs, buttons, stairs, etc)
+     */
+    public boolean renderAsNormalBlock()
+    {
+        return false;
+    }
+
+    /**
+     * The type of render function that is called for this block
+     */
+    public int getRenderType()
+    {
+        return 8;
+    }
+
+    /**
+     * Checks to see if its valid to put this block at the specified coordinates. Args: world, x, y, z
+     */
+    public boolean canPlaceBlockAt(World par1World, int par2, int par3, int par4)
+    {
+        return par1World.isBlockSolidOnSide(par2 - 1, par3, par4, EAST ) ||
+               par1World.isBlockSolidOnSide(par2 + 1, par3, par4, WEST ) ||
+               par1World.isBlockSolidOnSide(par2, par3, par4 - 1, SOUTH) ||
+               par1World.isBlockSolidOnSide(par2, par3, par4 + 1, NORTH);
+    }
+
+    public int func_85104_a(World par1World, int par2, int par3, int par4, int par5, float par6, float par7, float par8, int par9)
+    {
+        int var10 = par9;
+
+        if ((var10 == 0 || par5 == 2) && par1World.isBlockSolidOnSide(par2, par3, par4 + 1, NORTH))
+        {
+            var10 = 2;
+        }
+
+        if ((var10 == 0 || par5 == 3) && par1World.isBlockSolidOnSide(par2, par3, par4 - 1, SOUTH))
+        {
+            var10 = 3;
+        }
+
+        if ((var10 == 0 || par5 == 4) && par1World.isBlockSolidOnSide(par2 + 1, par3, par4, WEST))
+        {
+            var10 = 4;
+        }
+
+        if ((var10 == 0 || par5 == 5) && par1World.isBlockSolidOnSide(par2 - 1, par3, par4, EAST))
+        {
+            var10 = 5;
+        }
+
+        return var10;
+    }
+
+    /**
+     * Returns the quantity of items to drop on block destruction.
+     */
+    public int quantityDropped(Random par1Random)
+    {
+        return 1;
+    }
+
+    /**
+     * Returns the ID of the items to drop on destruction.
+     */
+    public int idDropped(int par1, Random par2Random, int par3)
+    {
+        return MinecraftPlusBase.holidaylights_2Idle.blockID;
+    }
+
+    /**
+     * only called by clickMiddleMouseButton , and passed to inventory.setCurrentItem (along with isCreative)
+     */
+    public int idPicked(World par1World, int par2, int par3, int par4)
+    {
+        return MinecraftPlusBase.holidaylights_2Idle.blockID;
+    }
+    
+}