+++ Source/minecraftplus_mod/BlockiBlock.java
@@ -46,7 +46,7 @@ public class BlockiBlock extends BlockNormal
              
         	 entityplayer.dropPlayerItem(new ItemStack(MinecraftPlusBase.embroniumIngot, 5));
          }
-
+         
          return false;
     }
 }